+++ package.json
@@ -28,7 +28,7 @@
   },
   "private": false,
   "homepage": "./",
-  "main": "public/electron.js",
+  "main": "build/electron.js",
   "scripts": {
     "start": "cross-env-shell NODE_ENV=development craco start",
     "start-electron": "concurrently \"npm run start\" \"wait-on http://localhost:3000 && electron .\"",
@@ -39,9 +39,10 @@
     "eject": "craco eject",
     "build-electron:portable": "cross-env-shell NODE_ENV=production REACT_APP_RELEASE_TYPE=portable ./node_modules/.bin/webpack --config ./scripts/electronWebpackConfig.js --colors && npm run moveBuildEntryPoint",
     "build-electron:setup": "cross-env-shell NODE_ENV=production REACT_APP_RELEASE_TYPE=setup ./node_modules/.bin/webpack --config ./scripts/electronWebpackConfig.js --colors && npm run moveBuildEntryPoint",
+    "build-electron:dev": "cross-env-shell NODE_ENV=development REACT_APP_RELEASE_TYPE=setup ./node_modules/.bin/webpack --config ./scripts/electronWebpackConfig.js --colors && npm run moveBuildEntryPoint",
     "moveBuildEntryPoint": "node ./scripts/moveNativesToBuild.js",
     "lint": "eslint .",
-    "dev": "cross-env-shell NODE_ENV=development APP_TYPE=electron BROWSER=none \"npm run start-electron\"",
+    "dev": "npm run build-electron:dev && cross-env-shell NODE_ENV=development APP_TYPE=electron BROWSER=none \"npm run start-electron\"",
     "dev-web": "cross-env-shell NODE_ENV=development APP_TYPE=web npm run start",
     "postinstall": "npm run rebuild",
     "rebuild": "node ./scripts/buildNativeDeps.js",
