+++ src/common/messageEvents.js
@@ -1,0 +1,11 @@
+const EV = {
+  // DISCORD RPC
+  INIT_DISCORD_RPC: 'INIT_DISCORD_RPC',
+  UPDATE_DISCORD_RPC: 'UPDATE_DISCORD_RPC',
+  SHUTDOWN_DISCORD_RPC: 'SHUTDOWN_DISCORD_RPC',
+
+  UPDATE_MOD_SYNC: 'UPDATE_MOD_SYNC',
+  UPDATE_MOD_SYNC_LONG: 'UPDATE_MOD_SYNC_LONG'
+};
+
+export default EV;
