+++ src/main/src/handleGlobalCrash.js
@@ -1,0 +1,6 @@
+import log from 'electron-log';
+
+process.on('uncaughtException', error => {
+  log.error(error);
+  process.exit();
+});
