+++ src/main/src/handleLegacyInstalls.js
@@ -1,0 +1,4 @@
+// Handle installs still using gdlauncher_next as appdata folder.
+// Move config and everything to gdlauncher and delete gdlauncher_next
+
+// Move override.data to userDataOverride.json
