+++ src/main/src/mainWindow.js
@@ -1,0 +1,122 @@
+import path from 'path';
+import {
+  shell,
+  BrowserWindow,
+  globalShortcut,
+  nativeImage,
+  Tray,
+  Menu
+} from 'electron';
+
+const isDev = process.env.NODE_ENV === 'development';
+
+// eslint-disable-next-line
+export let mainWindow;
+let tray;
+
+export const createWindow = () => {
+  mainWindow = new BrowserWindow({
+    width: 1100,
+    height: 700,
+    minWidth: 1100,
+    minHeight: 700,
+    show: true,
+    frame: false,
+    backgroundColor: '#1B2533',
+    webPreferences: {
+      experimentalFeatures: true,
+      nodeIntegration: true,
+      // Disable in dev since I think hot reload is messing with it
+      webSecurity: !isDev
+    }
+  });
+
+  if (isDev) {
+    globalShortcut.register('CommandOrControl+R', () => {
+      mainWindow.reload();
+    });
+
+    globalShortcut.register('F5', () => {
+      mainWindow.reload();
+    });
+  }
+
+  mainWindow.webContents.session.webRequest.onHeadersReceived(
+    {
+      urls: ['http://*/*', 'https://*/*']
+    },
+    (details, callback) => {
+      // eslint-disable-next-line
+      delete details.responseHeaders['Access-Control-Allow-Origin'];
+      // eslint-disable-next-line
+      delete details.responseHeaders['access-control-allow-origin'];
+      if (details.url.includes('www.google-analytics.com')) {
+        // eslint-disable-next-line
+        details.responseHeaders['Access-Control-Allow-Origin'] = [
+          'http://localhost:3000'
+        ];
+      } else {
+        // eslint-disable-next-line
+        details.responseHeaders['Access-Control-Allow-Origin'] = ['*'];
+      }
+      callback({
+        cancel: false,
+        responseHeaders: details.responseHeaders
+      });
+    }
+  );
+
+  const RESOURCE_DIR = isDev ? __dirname : path.join(__dirname, '../build');
+
+  const iconPath = path.join(RESOURCE_DIR, 'logo_32x32.png');
+
+  const nimage = nativeImage.createFromPath(iconPath);
+
+  tray = new Tray(nimage);
+  const trayMenuTemplate = [
+    {
+      label: 'GDLauncher',
+      enabled: false
+    },
+    {
+      label: 'Show Dev Tools',
+      click: () => mainWindow.webContents.openDevTools()
+    }
+  ];
+
+  const trayMenu = Menu.buildFromTemplate(trayMenuTemplate);
+  tray.setContextMenu(trayMenu);
+  tray.setToolTip('GDLauncher');
+  tray.on('double-click', () => mainWindow.show());
+
+  mainWindow.loadURL(
+    isDev
+      ? 'http://localhost:3000'
+      : `file://${path.join(__dirname, '../build/index.html')}`,
+    {
+      userAgent: 'GDLauncher'
+    }
+  );
+  if (isDev) {
+    mainWindow.webContents.openDevTools();
+  }
+
+  mainWindow.on('maximize', () => {
+    mainWindow.webContents.send('window-maximized');
+  });
+
+  mainWindow.on('unmaximize', () => {
+    mainWindow.webContents.send('window-minimized');
+  });
+
+  const handleRedirect = (e, url) => {
+    if (url !== mainWindow.webContents.getURL()) {
+      e.preventDefault();
+      shell.openExternal(url);
+    }
+  };
+
+  mainWindow.webContents.on('will-navigate', handleRedirect);
+  mainWindow.webContents.on('new-window', handleRedirect);
+  return mainWindow;
+};
